feat: migrate to a new files structure
